requirements.txt
pdfplumber
python-docx
beautifulsoup4
requests